\name{lambrusco}
\alias{lambrusco}
\docType{data}
\title{Wine data}
\description{
  A data frame \code{X} of LC-MS measurements of 76 lambrusco wines,
  from three different origins (indicated in the \code{sample.labels}
  variable). Only variables containing more than 10 non-zero values have
  been retained, bringing the total number of variables to 1,155.
}
\usage{data(lambrusco)}
\source{Personal communication, Georg Weingart}
\keyword{datasets}
