\name{runEx}
\alias{runEx}
\title{Run StatsDemo examples through shiny}
\description{Open the available examples in a shiny interface.
}
\usage{runEx(example)}
\arguments{
  \item{example}{Name of the example: a string. If no name is given all
    available examples will be shown.}
}
\author{Ron Wehrens}
\examples{
runEx()
\dontrun{
  runEx("PCA")
}
}
\keyword{environment}
